Use longs instead of ints for IEnergyStorage

+++ src/main/java/net/minecraftforge/energy/IEnergyStorage.java
@@ -39,7 +39,7 @@ public interface IEnergyStorage
     *            If TRUE, the insertion will only be simulated.
     * @return Amount of energy that was (or would have been, if simulated) accepted by the storage.
     */
-    int receiveEnergy(int maxReceive, boolean simulate);
+    long receiveEnergy(long maxReceive, boolean simulate);
 
     /**
     * Removes energy from the storage. Returns quantity of energy that was removed.
@@ -50,17 +50,17 @@ public interface IEnergyStorage
     *            If TRUE, the extraction will only be simulated.
     * @return Amount of energy that was (or would have been, if simulated) extracted from the storage.
     */
-    int extractEnergy(int maxExtract, boolean simulate);
+    long extractEnergy(long maxExtract, boolean simulate);
 
     /**
     * Returns the amount of energy currently stored.
     */
-    int getEnergyStored();
+    long getEnergyStored();
 
     /**
     * Returns the maximum amount of energy that can be stored.
     */
-    int getMaxEnergyStored();
+    long getMaxEnergyStored();
 
     /**
      * Returns if this storage can have energy extracted.
